# Optimize _plain_bfs functions