--- a/networkx/algorithms/components/connected.py
+++ b/networkx/algorithms/components/connected.py
@@ -188,7 +188,7 @@ def node_connected_component(G, n):
 
 def _plain_bfs(G, source):
     """A fast BFS node generator"""
-    adj = G.adj
+    adj = G._adj
     n = len(adj)
     seen = {source}
     nextlevel = [source]