@Tortar 2023-01-06
Implemented the same thing for the other `_plain_bfs` function for undirected graphs in `connected.py`, the corresponding benchmark table is

```
+--------------------------------------------------------+--------------------+---------------+--------------+
|                         graph                          | time: current main | time: this pr | times faster |
+--------------------------------------------------------+--------------------+---------------+--------------+
|     nx.random_regular_graph(d=1, n=10, seed=4241)      |        0.0         |      0.0      |     1.17     |
|     nx.random_regular_graph(d=9, n=10, seed=4241)      |        0.01        |      0.0      |     2.67     |
|    nx.random_regular_graph(d=10, n=100, seed=4241)     |        0.09        |      0.05     |     1.71     |
|    nx.random_regular_graph(d=99, n=100, seed=4241)     |        0.17        |      0.02     |     9.97     |
|    nx.random_regular_graph(d=1, n=1000, seed=4241)     |        0.0         |      0.0      |     1.17     |
|    nx.random_regular_graph(d=3, n=1000, seed=4241)     |        0.84        |      0.72     |     1.16     |
|    nx.random_regular_graph(d=10, n=1000, seed=4241)    |        0.86        |      0.65     |     1.31     |
|    nx.random_regular_graph(d=50, n=1000, seed=4241)    |        1.38        |      0.5      |     2.75     |
|   nx.random_regular_graph(d=100, n=1000, seed=4241)    |        1.94        |      0.34     |     5.76     |
|   nx.random_regular_graph(d=200, n=1000, seed=4241)    |        2.37        |      0.31     |     7.65     |
|   nx.random_regular_graph(d=500, n=1000, seed=4241)    |        5.1         |      0.26     |    19.43     |
|   nx.random_regular_graph(d=700, n=1000, seed=4241)    |        6.97        |      0.23     |    30.47     |
|     nx.erdos_renyi_graph(n=100, p=0.9, seed=4241)      |        0.12        |      0.02     |     7.26     |
|     nx.erdos_renyi_graph(n=1000, p=0.8, seed=4241)     |        7.34        |      0.21     |    35.41     |
| nx.watts_strogatz_graph(n=100, k=10, p=0.3, seed=4241) |        0.06        |      0.04     |     1.43     |
|           nx.connected_caveman_graph(100,10)           |        0.63        |      0.74     |     0.85     |
|                nx.caveman_graph(100,10)                |        0.01        |      0.01     |     0.85     |
|               nx.windmill_graph(10, 10)                |        0.05        |      0.02     |     2.87     |
|               nx.full_rary_tree(3, 1000)               |        0.48        |      0.33     |     1.47     |
|                 nx.complete_graph(100)                 |        0.13        |      0.01     |    10.34     |
|                 nx.complete_graph(10)                  |        0.01        |      0.0      |     1.9      |
|                  nx.cycle_graph(100)                   |        0.06        |      0.05     |     1.1      |
|                   nx.cycle_graph(10)                   |        0.01        |      0.01     |     1.21     |
+--------------------------------------------------------+--------------------+---------------+--------------+
```