@dschult 2023-01-09
This might be better with inputs `_plain_bfs(adj, source)` rather than `(G, source)`. That way functions that need to search a directed graph backwards can use it by inputing `G.pred`.  It does mean each place we call it we'll have to use update the call from using `G`.  

I'd be interested in seeing a comparison between calling it with `G.adj` and calling it with `G._adj`. `G.adj` is a view on `G._adj` so is perhaps slower, and since this function never updates G there is no risk of exposing the internal data structure to corruption.  (See #4895)