@Tortar 2023-01-09
Regarding your first observation, but then we should run two different bfs with the two inputs (G.pred, G.succ) right? This could have some performance degradation effects, what do you think?

Regarding `G._adj` I think that the comparison should be more or less the same as the one in #6299, so it should be very good to call `G._adj`! Is there something equivalent for `G.pred`?