@dschult 2023-01-09
I was wrong --- These functions are already restricted to being undirected (the one in connected.py) and directed (the one in weakly_connected.py).  So we should just input G as we currently do. As you point out, we'd have to call it twice (or with a `UnionAtlasView` of both G.pred and G.succ) and this is a better interface.

Looking at the `DiGraph.__init__` method you'll see that G.succ and G.pred are views of the dicts G._succ and G._pred.  So, to implementing this idea of using `G._adj` internally will amount to replacing `G.adj[u]` or `G[u]` with `G._adj[u]` and similarly replacing `G.succ[u]` with `G._succ[u]` and `G.pred[u]` with `G._pred[u]`.

The views are nice because they protect the user from messing up the data structure, but they are slower.