@Tortar 2023-01-10
oh didn't know about that (actually I'm new to Newtworkx internals :-)). Anyway I updated all the PRs reflecting your suggestions, I also updated the tables, all show other boost in performance!