@boothby 2023-05-27
Agreed, I applaud the effort to improve the performance of these functions in specific.  They form the backbone of so many algorithms throughout the library; a little elbow grease here goes a long way!